WIP: Settings refactor, moved everything in 2 folders
